# backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Motionclash
  version: 1.0.0
servers:
- url: https://localhost:7000
paths:
...
  /api/images:
    get:
      description: Returns a list of images in the database
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Image'
                type: array
          description: A list of images
      summary: Get a list of images
      tags:
      - Image
      x-accepts: application/json
      x-tags:
      - tag: Image
    post:
      description: Uploads a new image to the database
      requestBody:
        content:
          image/jpeg:
            schema:
              format: binary
              type: string
          image/png:
            schema:
              format: binary
              type: string
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
          description: The uploaded image
      summary: Upload a new image
      tags:
      - Image
      x-content-type: image/jpeg
      x-accepts: application/json
      x-tags:
      - tag: Image
  /api/images/{id}:
    delete:
      description: Deletes an image with the specified ID
      parameters:
      - description: The ID of the image to delete
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: Image deleted
        "404":
          description: Image not found
      summary: Delete an image by ID
      tags:
      - Image
      x-accepts: application/json
      x-tags:
      - tag: Image
    get:
      description: Returns an image with the specified ID
      parameters:
      - description: The ID of the image to retrieve
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application / json:
              schema:
                $ref: '#/components/schemas/Image'
          description: The requested image
        "404":
          description: Image not found
      summary: Get an image by ID
      tags:
      - Image
      x-accepts: application / json
      x-tags:
      - tag: Image
    put:
      description: Updates an image with the specified ID
      parameters:
      - description: The ID of the image to update
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application / json:
            schema:
              $ref: '#/components/schemas/Image'
        required: true
      responses:
        "200":
          content:
            application / json:
              schema:
                $ref: '#/components/schemas/Image'
          description: The updated image
        "404":
          description: Image not found
      summary: Update an image by ID
      tags:
      - Image
      x-content-type: application / json
      x-accepts: application / json
      x-tags:
      - tag: Image
  /api/image/race/{id}:
    get:
      operationId: getImagesOfRace
      parameters:
      - description: Id Of the reace to get images of
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Image'
                type: array
          description: List of Score objects in a race
      summary: Get all Scores of a race.
      tags:
      - Image
      x-accepts: application/json
      x-tags:
      - tag: Image
components:
  schemas:
...
    Image:
      example:
        imagePath: imagePath
        id: 123e4567-e89b-12d3-a456-426614174000
        imageType: ProfilePicture
      properties:
        id:
          example: 123e4567-e89b-12d3-a456-426614174000
          type: string
        imageType:
          enum:
          - ProfilePicture
          - EventPicture
          type: string
        imagePath:
          type: string
      type: object
...